+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 ## [Unreleased]
 
 ### Added
-- [`order`]: add `orderBySplitPaths` option
+- [`order`]: add `orderByFullPathString` option
 
 ## [2.32.0] - 2025-06-20